+++ scrape_cli/__init__.py
@@ -4,7 +4,7 @@ scrape-cli - A command-line tool to extract HTML elements using XPath or CSS3 se
 
 from scrape_cli.scrape import main
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"
 __author__ = "Andrea Borruso"
 __author_email__ = "aborruso@gmail.com"